feat: slim base install with configs only, heavy deps in [full] extra
`pip install prime-rl` now installs just `prime_rl.configs.*` (5 deps:
loguru, pydantic, pydantic-config, tomli, tomli-w). The full training
stack — torch, vllm, transformers, wandb, ring-flash-attn, etc. — moves
into a new `[full]` optional-dependencies extra. `pip install
prime-rl[full]` reproduces today's install.

To make this work, two small refactors:

- Move `rgetattr`/`rsetattr` from `utils/utils.py` (which imports torch
  and wandb at module top) into `utils/config.py` (stdlib + pydantic
  only). They're trivial dotted-path attr helpers used by
  `configs/inference.py` to translate our nested config into vLLM's
  flat argparse Namespace.
- Wrap the `_compat` import in `prime_rl/__init__.py` with
  try/except ImportError so the slim install (no transformers) doesn't
  fail on package import. Patches still run identically in the full
  install.

Verified end-to-end: a fresh venv with `pip install
/path/to/prime-rl` installs 13 packages and `from prime_rl.configs.rl
import RLConfig` works without torch/transformers/vllm/wandb/etc.
loaded.
