+++ pyproject.toml
@@ -5,16 +5,36 @@ description = "Async RL Training at Scale"
 readme = "README.md"
 requires-python = "~=3.12.0"
 dependencies = [
-    "beartype>=0.21.0",
-    "datasets>=4.0.0",
-    "jaxtyping>=0.3.2",
-    "liger-kernel>=0.5.10",
+    # Slim base: only what `prime_rl.configs.*` needs.
+    # Heavy runtime deps live in the `[full]` extra (or per-component extras).
     "loguru>=0.7.3",
-    "pyarrow>=21.0.0",
     "pydantic>=1.10.13",
     "pydantic-config",
     "tomli>=2.2.1",
     "tomli-w>=1.2.0",
+]
+
+[project.scripts]
+rl = "prime_rl.entrypoints.rl:main"
+sft = "prime_rl.entrypoints.sft:main"
+inference = "prime_rl.entrypoints.inference:main"
+trainer = "prime_rl.trainer.rl.train:main"
+orchestrator = "prime_rl.orchestrator.orchestrator:main"
+env-server = "prime_rl.orchestrator.env_server.env_server:main"
+
+[project.entry-points."vllm.general_plugins"]
+prime_rl = "prime_rl.inference.patches:transformers_v5_compat"
+
+[project.optional-dependencies]
+# Heavy runtime deps required to actually run trainer/orchestrator/inference.
+# `pip install prime-rl` gives you configs only; `pip install prime-rl[full]`
+# gives you the full training stack (today's default behavior).
+full = [
+    "beartype>=0.21.0",
+    "datasets>=4.0.0",
+    "jaxtyping>=0.3.2",
+    "liger-kernel>=0.5.10",
+    "pyarrow>=21.0.0",
     "numpy>=2.2.6",
     "torch>=2.9.0",
     "torchdata>=0.11.0",
@@ -37,19 +57,6 @@ dependencies = [
     "flash-linear-attention",
     "nvidia-ml-py>=12.575.51",
 ]
-
-[project.scripts]
-rl = "prime_rl.entrypoints.rl:main"
-sft = "prime_rl.entrypoints.sft:main"
-inference = "prime_rl.entrypoints.inference:main"
-trainer = "prime_rl.trainer.rl.train:main"
-orchestrator = "prime_rl.orchestrator.orchestrator:main"
-env-server = "prime_rl.orchestrator.env_server.env_server:main"
-
-[project.entry-points."vllm.general_plugins"]
-prime_rl = "prime_rl.inference.patches:transformers_v5_compat"
-
-[project.optional-dependencies]
 flash-attn = [
     "flash-attn @ https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.8.3+cu128torch2.10-cp312-cp312-linux_x86_64.whl ; platform_machine == 'x86_64'",
 ]
@@ -88,6 +95,7 @@ quack = [
     "quack-kernels>=0.3.3",
 ]
 all = [
+    "prime-rl[full]",
     "prime-rl[flash-attn]",
     "prime-rl[disagg]",
     "prime-rl[quack]",
